Widen getHistoryAtTimeStep to accept long index (#793)
Change Variable.getHistoryAtTimeStep and Flow.getHistoryAtTimeStep from
int to long parameter. Remove unsafe long-to-int cast in
Flows.pipelineDelay. Indices beyond ArrayList range safely return 0
instead of silently wrapping to negative values.

+++ courant-engine/src/main/java/systems/courant/sd/model/Flows.java
@@ -105,7 +105,7 @@ public final class Flows {
                                      LongSupplier stepSupplier, long delay) {
         return Flow.create(name, timeUnit, () -> {
             long referenceStep = stepSupplier.getAsLong() - delay;
-            double value = inflow.getHistoryAtTimeStep((int) referenceStep);
+            double value = inflow.getHistoryAtTimeStep(referenceStep);
             Stock sink = inflow.getSink();
             Preconditions.checkArgument(sink != null, "inflow must have a sink stock assigned");
             return new Quantity(value, sink.getUnit());
